# backend/openapi.yaml
openapi: 3.0.3
info:
  title: WindChaserSecurity API
  description: |
    API для образовательной платформы по информационной безопасности WindChaserSecurity.
    
    Платформа включает:
    - Управление пользователями и аутентификацию
    - Образовательные модули и задания
    - Отслеживание прогресса обучения
    - Чат с AI-ассистентами
  version: 1.0.0
  contact:
    name: WindChaserSecurity Team
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.windchaser.security
    description: Production server
...
paths:
...
  /api/me:
    get:
      tags: [Auth]
      summary: Получить текущего пользователя
      description: Возвращает информацию о текущем аутентифицированном пользователе
      operationId: getCurrentUser
      responses:
        '200':
          description: Успешно
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithRole'
        '401':
          description: Не аутентифицирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/login:
    post:
      tags: [Auth]
      summary: Вход в систему
      description: Аутентификация пользователя по ID
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithRole'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== Users ====================
...
  /api/tasks:
    get:
      tags: [Tasks]
      summary: Получить задания
      description: Возвращает все задания или фильтрует по module_id и flow_id
      operationId: getTasks
      parameters:
        - name: module_id
          in: query
          required: false
          description: ID модуля для фильтрации
          schema:
            type: string
            format: uuid
        - name: flow_id
          in: query
          required: false
          description: ID потока для дополнительной фильтрации
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Список заданий
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

    post:
      tags: [Tasks]
      summary: Создать задание
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Задание создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /api/tasks/{id}:
    parameters:
      - $ref: '#/components/parameters/TaskId'

    get:
      tags: [Tasks]
      summary: Получить задание по ID
      operationId: getTaskById
      responses:
        '200':
          description: Задание найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Задание не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags: [Tasks]
      summary: Обновить задание
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Задание обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Задание не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Tasks]
      summary: Удалить задание
      operationId: deleteTask
      responses:
        '204':
          description: Задание удалено
        '404':
          description: Задание не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== Progress ====================
  /api/user_task_progress:
    get:
      tags: [Progress]
      summary: Получить прогресс пользователя
      operationId: getUserProgress
      parameters:
        - name: user_id
          in: query
          required: true
          description: ID пользователя
          schema:
            type: string
            format: uuid
        - name: module_id
          in: query
          required: false
          description: ID модуля для фильтрации
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Прогресс пользователя
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserTaskProgress'
        '400':
          description: user_id обязателен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Progress]
      summary: Создать запись прогресса
      operationId: createProgress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProgressRequest'
      responses:
        '201':
          description: Прогресс создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserTaskProgress'
  /api/user_task_progress/{id}:
    parameters:
      - $ref: '#/components/parameters/ProgressId'

    patch:
      tags: [Progress]
      summary: Обновить прогресс
      operationId: updateProgress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProgressRequest'
      responses:
        '200':
          description: Прогресс обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserTaskProgress'
        '404':
          description: Запись прогресса не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== Chat Sessions ====================
...
components:
...
  parameters:
...
    TaskId:
      name: id
      in: path
      required: true
      description: ID задания
      schema:
        type: string
        format: uuid
    ProgressId:
      name: id
      in: path
      required: true
      description: ID записи прогресса
      schema:
        type: string
        format: uuid
...
  schemas:
...
    BaseEntity:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор
        created_at:
          type: string
          format: date-time
          description: Дата создания
        updated_at:
          type: string
          format: date-time
          description: Дата последнего обновления
        deleted_at:
          type: string
          format: date-time
          nullable: true
          description: Дата удаления (soft delete)

    # ---------- Error ----------
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Сообщение об ошибке
      example:
        message: "Resource not found"

    # ---------- Role ----------
    Role:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          required:
            - name
          properties:
            name:
              type: string
              description: Название роли
              example: "admin"

    # ---------- User ----------
    User:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          required:
            - role_id
            - username
            - email
          properties:
            role_id:
              type: string
              format: uuid
              description: ID роли пользователя
            username:
              type: string
              description: Имя пользователя
              example: "john_doe"
            email:
              type: string
              format: email
              description: Email пользователя
              example: "john@example.com"
            langflow_user_id:
              type: string
              nullable: true
              description: ID пользователя в LangFlow
            langflow_folder_id:
              type: string
              nullable: true
              description: ID папки в LangFlow
    UserWithRole:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          required:
            - role
          properties:
            role:
              $ref: '#/components/schemas/Role'

    # ---------- Module ----------
...
    Task:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          required:
            - module_id
            - title
            - type
            - description
            - max_score
          properties:
            module_id:
              type: string
              format: uuid
              description: ID модуля
            flow_id:
              type: string
              format: uuid
              nullable: true
              description: ID потока
            title:
              type: string
              description: Название задания
              example: "Анализ уязвимостей"
            type:
              type: string
              description: Тип задания
              example: "practical"
            description:
              type: string
              description: Описание задания
            max_score:
              type: number
              description: Максимальный балл
              example: 100
            achievement_badge:
              type: string
              nullable: true
              description: Значок достижения
    CreateTaskRequest:
      type: object
      required:
        - module_id
        - title
        - type
        - description
        - max_score
      properties:
        module_id:
          type: string
          format: uuid
        flow_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        type:
          type: string
        description:
          type: string
        max_score:
          type: number
        achievement_badge:
          type: string
          nullable: true
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        description:
          type: string
        max_score:
          type: number
        achievement_badge:
          type: string
          nullable: true

    # ---------- UserTaskProgress ----------
    UserTaskProgress:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          required:
            - user_id
            - task_id
            - status
          properties:
            user_id:
              type: string
              format: uuid
              description: ID пользователя
            task_id:
              type: string
              format: uuid
              description: ID задания
            status:
              type: string
              enum:
                - not_started
                - in_progress
                - completed
                - failed
              description: Статус выполнения
            score:
              type: number
              nullable: true
              description: Набранный балл
            started_at:
              type: string
              format: date-time
              nullable: true
              description: Дата начала
            completed_at:
              type: string
              format: date-time
              nullable: true
              description: Дата завершения
    CreateProgressRequest:
      type: object
      required:
        - user_id
        - task_id
        - status
      properties:
        user_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - not_started
            - in_progress
            - completed
            - failed
        score:
          type: number
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
    UpdateProgressRequest:
      type: object
      properties:
        status:
          type: string
          enum:
            - not_started
            - in_progress
            - completed
            - failed
        score:
          type: number
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true

    # ---------- AssistantProfile ----------
...
    LoginRequest:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: string
          format: uuid
          description: ID пользователя для входа